rename labeling/ to labeler/
