+++ labeler/keyword_labeler_test.go
@@ -1,4 +1,4 @@
-package labeling
+package labeler
 
 import (
 	"fmt"
